# Conflicts:
#	mainProject/src/com/stockmarket/www/dao/jdbc/JdbcHaveStockDao.java

+++ mainProject/src/com/stockmarket/www/service/basic/BasicAssetTrendService.java
@@ -44,8 +44,9 @@ public class BasicAssetTrendService implements AssetTrendService {
 		for (HaveStockView data : list) {
 			// (보유종목당 현재가 및 보유수량 확인용)
 			// System.out.println(data.getPrice()+","+data.getQuantity()); 
-			int presentValue = Integer.parseInt(data.getPrice().replaceAll(",",""));
-			int quantity = data.getQuantity();
+			long presentValue = Long.parseLong(data.getPrice().replaceAll(",",""));
+			long quantity = data.getQuantity();
+			// System.out.println(data.getStockName()+": "+presentValue*quantity);
 			sum += presentValue*quantity;
 		}
 		// 가상머니+(현재가*보유수량)+(현재가*보유수량).....
